feat(rpc): return serde errors as jsonrpc responses
