+++ rpc/src/util/error_codes.rs
@@ -4,6 +4,7 @@ pub const NOT_SUPPORTED_CODE: i32 = -32002;
 pub const INVALID_BOC_CODE: i32 = -32003;
 pub const TOO_LARGE_LIMIT_CODE: i32 = -32004;
 pub const TIMEOUT_CODE: i32 = -32005;
+pub const BAD_RESPONSE_CODE: i32 = -32006;
 
 pub const PARSE_ERROR_CODE: i32 = -32700;
 pub const INVALID_REQUEST_CODE: i32 = -32600;
